refactor: entities list and chart view components moved to index component (#447)
