+++ src/components/Index/components/Hero/components/Summaries/stories/summaries.stories.tsx
@@ -1,19 +1,15 @@
 import { Meta, StoryObj } from "@storybook/react";
-import { Summaries } from "./summaries";
+import { Summaries } from "../summaries";
 
-const meta = {
-  argTypes: {
-    summaries: { control: "object" },
-  },
+const meta: Meta<typeof Summaries> = {
   component: Summaries,
-  title: "Components/Summary",
-} satisfies Meta<typeof Summaries>;
+};
 
 export default meta;
 
 type Story = StoryObj<typeof meta>;
 
-export const SummariesStory: Story = {
+export const Default: Story = {
   args: {
     summaries: [
       { count: "1", label: "Species" },
