+++ src/components/Index/index.tsx
@@ -1,5 +1,7 @@
 import React from "react";
 import { useLayoutDimensions } from "../../providers/layoutDimensions/hook";
+import { ChartView } from "./components/EntitiesView/components/ChartView/chartView";
+import { EntityList } from "./components/EntitiesView/components/EntityList/entityList";
 import { EntitiesView } from "./components/EntitiesView/entitiesView";
 import { useEntitiesView } from "./components/EntitiesView/hooks/UseEntitiesView/hook";
 import { VIEW_MODE } from "./components/EntitiesView/hooks/UseEntitiesView/types";
@@ -8,10 +10,12 @@ import { Index as IndexLayout } from "./index.styles";
 import { IndexProps } from "./types";
 
 export const Index = ({
-  chart,
+  categoryFilters,
   className,
-  list,
+  entityListType,
+  entityName,
   ListHero,
+  loading,
   SideBarButton,
   SubTitleHero,
   Summaries,
@@ -31,7 +35,15 @@ export const Index = ({
         viewMode={viewMode}
         viewStatus={viewStatus}
       >
-        {viewMode === VIEW_MODE.TABLE ? list : chart}
+        {viewMode === VIEW_MODE.TABLE ? (
+          <EntityList entityListType={entityListType} />
+        ) : (
+          <ChartView
+            categoryFilters={categoryFilters}
+            entityName={entityName}
+            loading={loading}
+          />
+        )}
       </EntitiesView>
     </IndexLayout>
   );
