+++ src/components/Index/types.ts
@@ -1,10 +1,14 @@
 import { BaseComponentProps } from "components/types";
+import { EntityConfig } from "config/entities";
 import { ReactNode } from "react";
+import { CategoryFilter } from "../../components/Filter/components/Filters/filters";
 
 export interface IndexProps extends BaseComponentProps {
-  chart?: ReactNode;
-  list?: ReactNode;
+  categoryFilters: CategoryFilter[];
+  entityListType: string;
+  entityName: EntityConfig["label"];
   ListHero?: ReactNode | ReactNode[];
+  loading: boolean;
   SideBarButton?: ReactNode;
   SubTitleHero?: ReactNode | ReactNode[];
   Summaries?: ReactNode;
